requirements.txt
streamlit==1.28.1
google-generativeai==0.4.0
python-dotenv==1.0.0
pandas==2.1.3
plotly==5.17.0
wordcloud==1.9.2
